Makefile.mingw: Force GCC to tune for generic CPUs on x86_64.
This is needed in case we use the 64-bit version of NSXfer in the future.
We can't guarantee that the user's toolchain hasn't been built in a way
that causes GCC to tune for a specific CPU (or a later functional level),
so let's ensure that we're using generic instructions no matter what.